kubepkg: Move debs subcommand to separate cmd directory
