+++ cmd/kubepkg/cmd/BUILD.bazel
@@ -1,0 +1,30 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "debs.go",
+        "root.go",
+    ],
+    importpath = "k8s.io/release/cmd/kubepkg/cmd",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@com_github_blang_semver//:go_default_library",
+        "@com_github_google_go_github_v28//github:go_default_library",
+        "@com_github_spf13_cobra//:go_default_library",
+    ],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
